Rename Dry::Transaction::Sequence to Dry::Transaction
This class was originally CallSheet::Transaction, and we had to find a new name when we renamed the gem to dry-transaction. Sequence never felt like the right name, especially when the object we’re meant to be modelling here is a _business transaction_. This change moves back to using that name directly.
