+++ lib/dry/transaction/version.rb
@@ -1,6 +1,6 @@
 module Dry
   # Business transaction DSL.
-  module Transaction
+  class Transaction
     VERSION = "0.8.0".freeze
   end
 end
